requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.3
idna==3.7
pillow==10.2.0
pyasn1==0.6.0
pycryptodome==3.20.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
